requirements.txt
Django==3.2.5
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
Pillow==9.1.0
django-etc==1.3.1
django-cleanup==6.0.0
whitenoise==6.0.0